crawler/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
pydantic==2.4.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
locust==2.20.0
Pillow==10.2.0
torch==2.2.0
transformers==4.37.2
numpy==1.26.4
sqlalchemy==2.0.29
alembic==1.13.1
prometheus_client==0.20.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
stripe==8.9.0
jinja2==3.1.2
requests==2.31.0
jmespath==1.0.1
psycopg2-binary==2.9.9 